simulator/requirements.txt
azure-core==1.35.1
azure-eventhub==5.15.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
